-split check major changes -android icons fixed

+++ lib/providers/split_provider.dart
@@ -1,4 +1,4 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-final requiredAmountProvider = StateProvider<double>((ref) => 0.0);
+final requiredAmountProvider = StateProvider<int>((ref) => 0);
 final amountPerBatchProvider = StateProvider<int>((ref) => 0);
